refactor: update random sampling methods to improve type stability
- Changed parameter types in various random sampling methods from `usize` to `usize` for consistency and clarity.
- Updated method signatures in `simulate_fbm`, `simulate_levy_walk_with_step`, `simulate_asymmetric_levy`, and other related functions to reflect the new types.
- Removed unnecessary conversions in method bodies, enhancing readability and maintainability across the codebase.

+++ src/simulation/continuous/fbm.rs
@@ -128,7 +128,7 @@ pub fn simulate_fbm(
 
     // Generate fBn
     let noise = if hurst_exponent == 0.5 {
-        normal::rands(0.0, 2.0 * time_step, num_steps as usize)?
+        normal::rands(0.0, 2.0 * time_step, num_steps)?
     } else {
         circulant.generate()?
     };
